jnlblkblkjb

+++ VAST/core/submodels/output_submodels/vlm_post_processing/horseshoe_circulations.py
@@ -80,7 +80,7 @@ class HorseshoeCirculations(Model):
         mtx_val[rows, cols] = data
 
 
-        A = mtx = self.create_input('mtx', val=mtx_val)
+        A = mtx = self.create_input('mtx', val=mtx_val.astype(int))
 
         # gamma_b = self.declare_variable(
         #     'gamma_b', shape_by_conn=True)  # shape_by_conn not working
